requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
asyncio==3.4.3
attrs==22.1.0
beautifulsoup4==4.12.2
black==22.3.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.1
frozenlist==1.3.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
multidict==6.0.2
mypy-extensions==0.4.3
newrelic==8.10.1
pathspec==0.9.0
platformdirs==2.5.2
python-decouple==3.6
requests==2.27.1
soupsieve==2.4.1
structlog==23.1.0
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
Wand==0.6.7
Werkzeug==2.1.1
yarl==1.8.1
